requirements.txt
asgiref==3.5.0
autopep8==1.6.0
cycler==0.11.0
Django==4.0.4
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.31.1
kiwisolver==1.4.0
matplotlib==3.5.1
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
pygame==2.1.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
